requirements.txt
amqp==1.0.13
anyjson==0.3.3
asgiref==3.3.4
billiard==3.6.4.0
celery==3.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-import-export==2.5.0
dnspython==2.1.0
et-xmlfile==1.1.0
greenlet==1.1.1
kombu==2.5.16
MarkupPy==1.14
numpy==1.21.2
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.3
Pillow==8.2.0
prompt-toolkit==3.0.20
psycopg2-binary==2.8.6
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
